+++ CHANGELOG.md
@@ -27,11 +27,10 @@
 
 **NOTE**: For a smooth upgrade path from `0.21` to `> 0.22`
 on an existing deployment, this version must not be skipped
-or the provided feature flags used!
+or the provided legacy configuration for `libp2p-noise` used!
 
 - Bump `libp2p-noise` dependency to `0.22`. See the `libp2p-noise`
-changelog for details. The corresponding feature flag `noise-send-legacy-handshake`
-is provided through the `libp2p` crate.
+changelog for details about the `LegacyConfig`.
 
 - Refactored bandwidth logging ([PR 1670](https://github.com/libp2p/rust-libp2p/pull/1670)).