+++ CHANGELOG.md
@@ -25,6 +25,13 @@
 
 # Version 0.23.0 (2020-??-??)
 
+**NOTE**: For a smooth upgrade path from `0.21` to `> 0.22`
+on an existing deployment, this version must not be skipped
+or the provided legacy configuration for `libp2p-noise` used!
+
+- Bump `libp2p-noise` dependency to `0.22`. See the `libp2p-noise`
+changelog for details about the `LegacyConfig`.
+
 - Refactored bandwidth logging ([PR 1670](https://github.com/libp2p/rust-libp2p/pull/1670)).
 
 # Version 0.22.0 (2020-07-17)